refactor and add file swals
