+++ src/App.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import SideBar from '../src/Components/SideBar/SideBar'
 import NavBar from './Components/NavBar/NavBar'
-import FileHistory from './Components/FileTable/FileHistory'
+import FileHistory from './Components/Sarthak/FileTable/FileHistory'
 import './App.css'
-import FileTable from './Components/FileTable/FileTable'
+import FileTable from './Components/Sarthak/FileTable/FileTable'
 import ProfileCard from './Components/ProfileCard/ProfileCard'
 import { Routes, Route } from 'react-router-dom'
 import MyFiles from './Pages/MyFiles'
